+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.3.1] 2022-10-31
+### Changed
+- AudioCaps, Clotho and MACS order are now defined by their order in the corresponding captions CSV files when available.
+- Update documentation usage and main page.
+
+### Fixed
+- Workflow when requirements cache is invalid.
+
 ## [0.3.0] 2022-09-28
 ### Added
 - Add `column_names`, `info` and `shape` properties in datasets.